Update tub1001.cls
in the definition of the new environment "projeyonetimiekipvearastirmaolanaklari", there is a call to an undefined function bursiyerlistesi.
It ws probably defined so as to add a list of grantees. it is not necessary to add this list. since the function does not do anything, it created a valid proposal issuing an error warning. 
Now, if that function is not defined, it is not called. Still, if in the proposal, one defines this function, it will be called and a list of grantees will be added to the proposal.